Changed runner.pl to allow you to specify a parser, all, or none.
runner now takes an extra parameter, which can be all, none, or the name of a parser. When a specific parser is specified, it and the google links parsers are the only ones that run.

In general, runner will be called twice for a collection - once as runner.pl none (which gets all the pages) and then once as runner.pl -s all.