Implemented Fixup
Fixup is a process that lets you download and parse the results for
just one website after a regular download and parse has been completed.
The fixup.pl program will take the location of the searchResults.txt
file, a parser to use, and output/data directories for downloaded
pages/result data.